go back to making these install constant subs
when (Foo) { ... } where Foo is a non-constant sub actually just checks
for the boolean result of Foo(), because given/when are really broken.
Constant folding avoids that, because given/when are even more broken.

This change at least makes given (Foo->new) { when (Foo) { ... } } work
again, hopefully in a way that's future-compatible with the new
smart-match proposal.